requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
cloudinary==1.36.0
Django==4.2.7
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
idna==3.4
Pillow==10.1.0
Pygments==2.16.1
python-dotenv==1.0.0
pytz==2023.3.post1
redgreenunittest==0.1.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
